+++ packages/eyes-testcafe/CHANGELOG.md
@@ -1,14 +1,13 @@
 # Change Log
 
-- set failTestcafeOnDiff default value to true
 ## Unreleased
+- set failTestcafeOnDiff default value to true
 - updated to @applitools/eyes-api@1.1.6 (from 1.1.0)
 - updated to @applitools/eyes-sdk-core@12.24.9 (from 12.23.3)
 - updated to @applitools/utils@1.2.11 (from 1.2.2)
 - updated to @applitools/visual-grid-client@15.8.55 (from 15.8.24)
 - updated to @applitools/eyes-sdk-core@12.24.13 (from 12.24.9)
 - updated to @applitools/visual-grid-client@15.8.60 (from 15.8.55)
-
 
 ## 1.15.0 - 2021/9/6